% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-bundle.R
\name{add_bundle_concept_relationship}
\alias{add_bundle_concept_relationship}
\title{Add bundle concept relationship}
\usage{
add_bundle_concept_relationship(
  bundle_concepts,
  bundle_id,
  concept_key,
  concept_row
)
}
\arguments{
\item{bundle_concepts}{Data frame. The bundle concepts data frame.}

\item{bundle_id}{Character. The bundle id.}

\item{concept_key}{Character. The concept key.}

\item{concept_row}{Data frame. The concept row.}
}
\value{
A data frame.
}
\description{
Adds a bundle concept relationship to the bundle concepts data frame.
}
\keyword{internal}
