% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundles.R
\name{available_bundles}
\alias{available_bundles}
\title{Get available bundles for a version}
\usage{
available_bundles(version = "latest")
}
\arguments{
\item{version}{Requested version, if not defined, the latest will be used}
}
\value{
dataframe that contains a "bundle_name", "version" and a "domain" column for each available bundle
}
\description{
\strong{Deprecated:} This function is deprecated. Use \code{\link{list_bundles}} instead.
}
\examples{
# Deprecated - use list_bundles() instead
# available_bundles()
# available_bundles("0.1")
}
