% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-bundle.R
\name{check_and_add_concept}
\alias{check_and_add_concept}
\title{Add concept if it doesn't already exist}
\usage{
check_and_add_concept(concepts_df, concept_key, concept_row)
}
\arguments{
\item{concepts_df}{Data frame. The concepts data frame.}

\item{concept_key}{Character. The concept key.}

\item{concept_row}{Data frame. The concept row.}
}
\value{
A data frame.
}
\description{
Checks if a concept already exists in the concepts data frame and adds it if it doesn't.
}
\keyword{internal}
