% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundles.R
\name{concept_by_bundle}
\alias{concept_by_bundle}
\title{Get concepts for a a single bundle row}
\usage{
concept_by_bundle(domain, id, version = "latest")
}
\arguments{
\item{domain}{The domain of the bundle.}

\item{id}{The ID of the bundle.}

\item{version}{The version of the bundle. Default is "latest".}
}
\value{
A data frame with the concept data.
}
\description{
\strong{Deprecated:} This function is deprecated. Use \code{\link{get_bundle_concepts}} instead.
}
\examples{
# Deprecated - use get_bundle_concepts() instead
# concept_by_bundle(domain = "observation", id = "smoking")
}
