% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-bundle.R
\name{delete_bundle_concepts_json}
\alias{delete_bundle_concepts_json}
\title{Delete bundle concepts JSON file}
\usage{
delete_bundle_concepts_json(bundle_id)
}
\arguments{
\item{bundle_id}{Character. Bundle ID}
}
\description{
Deletes the JSON file for a bundle.
}
\keyword{internal}
