% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-bundle-concepts.R
\name{expand_descendants}
\alias{expand_descendants}
\title{Expand concept descendants}
\usage{
expand_descendants(vocab_conn, concept_ids)
}
\arguments{
\item{vocab_conn}{Vocab connection object}

\item{concept_ids}{Integer vector. Concept IDs to expand}
}
\value{
Integer vector of all concept IDs (including descendants)
}
\description{
Expands a set of concept IDs to include all descendant concepts
  using the concept_ancestor table.
}
\keyword{internal}
