% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-bundle-concepts.R
\name{get_concept_metadata}
\alias{get_concept_metadata}
\title{Get concept metadata from vocabulary connection}
\usage{
get_concept_metadata(vocab_conn, concept_ids)
}
\arguments{
\item{vocab_conn}{Vocab connection object}

\item{concept_ids}{Integer vector. Concept IDs to retrieve}
}
\value{
Data frame with concept metadata
}
\description{
Retrieves concept metadata (name, domain, etc.) for a set of
  concept IDs.
}
\keyword{internal}
