% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocab-connection.R
\name{get_concept_table}
\alias{get_concept_table}
\title{Get concept table from vocabulary connection}
\usage{
get_concept_table(vocab_conn)
}
\arguments{
\item{vocab_conn}{Vocab connection object created by create_vocab_connection()}
}
\value{
Concept table (data frame or database table reference)
}
\description{
Retrieves the concept table from a vocabulary connection.
}
\keyword{internal}
