% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-dir.R
\name{get_data_raw_dir}
\alias{get_data_raw_dir}
\title{Get data-raw directory path}
\usage{
get_data_raw_dir()
}
\value{
Character string with path to data-raw directory
}
\description{
Returns the path to the data-raw directory. This function can
  be mocked in tests to use a temporary directory.
}
\keyword{internal}
