% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-bundle.R
\name{process_child_bundles}
\alias{process_child_bundles}
\title{Process child bundles}
\usage{
process_child_bundles(bundles, bundle_id, child_bundles, include_descendants)
}
\arguments{
\item{bundles}{Data frame. The bundles data frame.}

\item{bundle_id}{Character. The bundle id.}

\item{child_bundles}{Character vector. The child bundles to process.}

\item{include_descendants}{Logical. Whether to include descendants.}
}
\value{
A data frame.
}
\description{
Processes child bundles if provided.
}
\keyword{internal}
