% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-data.R
\name{raw_concept_by_bundle}
\alias{raw_concept_by_bundle}
\title{Get concepts for a single bundle row}
\usage{
raw_concept_by_bundle(domain, id)
}
\arguments{
\item{domain}{The domain of the bundle.}

\item{id}{The ID of the bundle.}
}
\value{
A data frame with the concept data.
}
\description{
Retrieves concept data for a specific bundle.
}
\keyword{internal}
