% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-bundle.R
\name{read_all_bundle_hierarchy_json}
\alias{read_all_bundle_hierarchy_json}
\title{Read all bundle hierarchy from JSON files}
\usage{
read_all_bundle_hierarchy_json()
}
\value{
Data frame with columns: parent_bundle_id, child_bundle_id, is_excluded
}
\description{
Reads all bundle hierarchy from JSON files and returns a data frame
  in the same format as the old CSV file (but without include_descendants).
}
\keyword{internal}
