% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-bundle.R
\name{read_raw_data}
\alias{read_raw_data}
\title{Read raw data from the data-raw directory}
\usage{
read_raw_data(table = c("bundles", "concepts"))
}
\arguments{
\item{table}{Character. The table to read.}
}
\value{
A data frame.
}
\description{
Reads raw data from the data-raw directory into a data frame.
}
\keyword{internal}
