% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-bundle-concepts.R
\name{resolve_bundle_hierarchy}
\alias{resolve_bundle_hierarchy}
\title{Resolve bundle hierarchy recursively}
\usage{
resolve_bundle_hierarchy(bundle_ids, bundle_hierarchy, bundles)
}
\arguments{
\item{bundle_ids}{Character vector. Bundle IDs to resolve}

\item{bundle_hierarchy}{Data frame. Bundle hierarchy relationships}

\item{bundles}{Data frame. All bundles}
}
\value{
Character vector of all bundle IDs (including children)
}
\description{
Recursively resolves all child bundles for a given set of
  bundle IDs.
}
\keyword{internal}
