% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocab-connection.R
\name{resolve_concept_code}
\alias{resolve_concept_code}
\title{Resolve concept_code to concept_id}
\usage{
resolve_concept_code(vocab_conn, concept_code, vocabulary_id)
}
\arguments{
\item{vocab_conn}{Vocab connection object}

\item{concept_code}{Character vector. Concept codes to resolve}

\item{vocabulary_id}{Character vector. Vocabulary IDs corresponding to concept codes}
}
\value{
Data frame with concept_id, concept_code, vocabulary_id
}
\description{
Resolves a concept_code and vocabulary_id to a concept_id using
  the vocabulary connection.
}
\keyword{internal}
