% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-bundle.R
\name{upsert_bundle_metadata}
\alias{upsert_bundle_metadata}
\title{Update or insert bundle metadata}
\usage{
upsert_bundle_metadata(
  bundles,
  bundle_exists,
  bundle_id,
  bundle_name,
  description
)
}
\arguments{
\item{bundles}{Data frame. The bundles data frame.}

\item{bundle_exists}{Logical. Whether the bundle exists.}

\item{bundle_id}{Character. The bundle id.}

\item{bundle_name}{Character. The bundle name to update or insert.}

\item{description}{Character. The bundle description to update or insert.}
}
\value{
A data frame.
}
\description{
Update or insert bundle metadata into the bundles data frame.
}
\keyword{internal}
