% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-bundle.R
\name{write_bundle_concepts_json}
\alias{write_bundle_concepts_json}
\title{Write bundle concepts and child bundles to JSON file}
\usage{
write_bundle_concepts_json(bundle_id, concepts_df, child_bundles_df = NULL)
}
\arguments{
\item{bundle_id}{Character. Bundle ID}

\item{concepts_df}{Data frame. Concepts data frame with columns: concept_key, include_descendants, is_excluded}

\item{child_bundles_df}{Data frame. Child bundles data frame with columns: child_bundle_id, is_excluded (optional)}
}
\description{
Writes concepts and child bundles for a bundle to its JSON file.
}
\keyword{internal}
