% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundles.R
\name{available_bundles}
\alias{available_bundles}
\title{Get available bundles for a version}
\usage{
available_bundles(version = "latest")
}
\arguments{
\item{version}{Requested version, if not defined, the latest will be used}
}
\value{
dataframe that contains a "bundle_name", "version" and a "domain" column for each available bundle
}
\description{
If a bundle has multiple names, then the id will be duplicated across rows
}
\examples{
available_bundles()
available_bundles("0.1")
}
