% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundles.R
\name{concept_by_bundle}
\alias{concept_by_bundle}
\title{Get concepts for a a single bundle row}
\usage{
concept_by_bundle(domain, id, version = "latest")
}
\arguments{
\item{domain}{The domain of the bundle.}

\item{id}{The ID of the bundle.}

\item{version}{The version of the bundle. Default is "latest".}
}
\value{
A data frame with the concept data.
}
\description{
Retrieves concept data for a specific bundle.
}
\examples{
# Using with the rest of the package
smoking_info <- available_bundles() |> dplyr::filter(concept_name == "smoking")
concept_by_bundle(domain = smoking_info$domain, id = smoking_info$id, version = smoking_info$version)
# Using if you know the details directly
concept_by_bundle(domain = "observation", id = "smoking.csv")
}
