% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundles.R
\name{concept_by_bundle}
\alias{concept_by_bundle}
\title{Get concepts for a a single bundle row}
\usage{
concept_by_bundle(bundle_row)
}
\arguments{
\item{bundle_row}{Single row of a dataframe with a domain and id}
}
\value{
Dataframe with a concept_id and domain column
}
\description{
Get concepts for a a single bundle row
}
\examples{
available_bundles() |> dplyr::filter(concept_name == "smoking") |> concept_by_bundle()
dplyr::tibble(id = "smoking", domain = "observation", version = "latest") |> concept_by_bundle()
}
