% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundles.R
\name{concept_by_bundle}
\alias{concept_by_bundle}
\title{Get concepts for a a single bundle row}
\usage{
concept_by_bundle(domain, id, version = "latest")
}
\arguments{
\item{domain}{The domain of the bundle.}

\item{id}{The ID of the bundle.}

\item{version}{The version of the bundle. Default is "latest".}
}
\value{
A data frame with the concept data.
}
\description{
Retrieves concept data for a specific bundle.
}
\examples{
# Usage with available_bundles, from a single row
smoking <- available_bundles() |> dplyr::filter(bundle_name == "Smoking")
concept_by_bundle(domain = smoking$domain, id = smoking$id, version = smoking$version)
# Using if you know the details directly
concept_by_bundle(domain = "observation", id = "smoking")
}
