% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocab-connection.R
\name{create_vocab_connection}
\alias{create_vocab_connection}
\title{Create vocabulary connection object}
\usage{
create_vocab_connection(connection, connection_type = c("OMOP", "SNOMED"))
}
\arguments{
\item{connection}{Connection object. Connection to vocabulary database
(database connection, Arrow dataset, or list of data frames)}

\item{connection_type}{Character. Type of connection: "OMOP" or "SNOMED".
Currently only "OMOP" is supported.}
}
\value{
Vocab connection object that encapsulates the connection and its type
}
\description{
Creates a vocabulary connection object that encapsulates the
  connection to an OMOP vocabulary database and its connection type.
}
\examples{
# CDM connection for OMOP
# cdm <- CDMConnector::cdmFromCon(con,
#   cdmName = "eunomia",
#   cdmSchema = "main",
#   writeSchema = "main"
# )
# vocab_conn <- create_vocab_connection(
#   connection = cdm,
#   connection_type = "OMOP"
# )

# Data frames for testing
# concept_df <- data.frame(...)
# ancestor_df <- data.frame(...)
# cdm <- list(
#   concept = concept_df,
#   concept_ancestor = ancestor_df
# )
# vocab_conn <- create_vocab_connection(
#   connection = cdm,
#   connection_type = "OMOP"
# )
}
