% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw-data.R
\name{raw_bundles}
\alias{raw_bundles}
\title{Get available raw bundles}
\usage{
raw_bundles()
}
\value{
dataframe that contains a "bundle_name" and a "domain" column for each available bundle
}
\description{
If a bundle has multiple names, then the id will be duplicated across rows
}
\keyword{internal}
